pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "argostranslate>=1.10.0",
    "async-timeout>=5.0.1",
    "beautifulsoup4>=4.14.2",
    "chromadb>=0.3.23",
    "faiss-cpu>=1.13.0",
    "ftfy>=6.3.1",
    "google-cloud-translate>=3.23.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.1",
    "langchain-openai>=1.0.3",
    "langdetect>=1.0.9",
    "lxml>=6.0.2",
    "pdfplumber>=0.11.8",
    "pillow>=12.0.0",
    "pypdf>=6.4.0",
    "pypdfium2>=5.0.0",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-slugify>=8.0.4",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.12.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "transformers>=4.57.1",
]
